requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
Flask-Mail==0.9.1
gunicorn==21.2.0